Revert "upload branch for auth (#73)"
This reverts commit df3c92577602c7559dfbfb4da58665c3d3b0aa1b.
