+++ app/package.json
@@ -14,8 +14,7 @@
   "dependencies": {
     "cors": "^2.8.5",
     "dotenv": "^16.4.5",
-    "express": "^4.21.2",
-    "express-openid-connect": "^2.17.1",
+    "express": "^4.20.0",
     "moment-timezone": "^0.5.45",
     "pg": "^8.12.0"
   },
