--- client/.env.example
@@ -1,2 +1,0 @@
-NEXT_PUBLIC_DOMAIN=
-NEXT_PUBLIC_CLIENTID=
