+++ client/package.json
@@ -11,8 +11,6 @@
     "test": "jest"
   },
   "dependencies": {
-    "@auth0/auth0-react": "^2.2.4",
-    "dotenv": "^16.4.7",
     "next": "^14.2.13",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
